removed orthohonality
git-svn-id: https://asn1c.svn.sourceforge.net/svnroot/asn1c/trunk@376 59561ff5-6e30-0410-9f3c-9617f08c8826

+++ libasn1parser/asn1p_expr_str.h
@@ -51,7 +51,7 @@ static char *asn1p_expr_type2str[] __attribute__ ((unused)) = {
 	|| ((size_t)(type)) >= sizeof(asn1p_expr_type2str)	\
 		/ sizeof(asn1p_expr_type2str[0]))		\
 		? (char *)0					\
-		: asn1p_expr_type2str[(type)]			\
+		: asn1p_expr_type2str[(int)(type)]		\
 	)
 
 #endif	/* ASN1_PARSER_EXPR_STR_H */
